requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
distro==1.9.0
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
groq==0.37.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
django
django-filter
djangorestframework
groq
gunicorn
numpy
pandas
pillow
scikit-learn
whitenoise
django
django-filter
djangorestframework
groq
gunicorn
numpy
pandas
pillow
scikit-learn
whitenoise
